projects/sentiment_stock_predictor/requirements.txt
praw
tweepy
yfinance
pandas
numpy
vaderSentiment
matplotlib
seaborn
pyyaml